LPS-174018 Prepare by moving to object-api
