+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/info/item/renderer/init.jsp
@@ -19,9 +19,9 @@
 <%@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %>
 
 <%@ page import="com.liferay.asset.display.page.portlet.AssetDisplayPageFriendlyURLProvider" %><%@
+page import="com.liferay.object.constants.ObjectWebKeys" %><%@
 page import="com.liferay.object.model.ObjectEntry" %><%@
 page import="com.liferay.object.rest.dto.v1_0.Link" %><%@
-page import="com.liferay.object.web.internal.constants.ObjectWebKeys" %><%@
 page import="com.liferay.portal.kernel.util.Validator" %>
 
 <%@ page import="java.io.Serializable" %>
